requirements.txt
astronomer-cosmos==1.11.0
dbt-core==1.8.0
dbt-postgres==1.8.0
psycopg2-binary==2.9.11
pandas==2.1.4
pyarrow==18.0.0
boto3==1.35.36
minio==7.2.18
openpyxl==3.1.2
dbt-adapters==1.4.0
